Enum testrunner::reporter::TestEvent

pub enum TestEvent<'a> {
    RunStarted {
        test_list: &'a TestList,
    },
    TestStarted {
        test_instance: TestInstance<'a>,
    },
    TestFinished {
        test_instance: TestInstance<'a>,
        run_status: TestRunStatus,
    },
    TestSkipped {
        test_instance: TestInstance<'a>,
    },
    RunBeginCancel {
        running: usize,
        reason: CancelReason,
    },
    RunFinished {
        start_time: Instant,
        initial_run_count: usize,
        run_stats: RunStats,
    },
}

Variants

RunStarted

The test run started.

Fields of RunStarted

test_list: &'a TestList

The list of tests that will be run.

The methods on the test list indicate the number of

TestStarted

A test started running.

Fields of TestStarted

test_instance: TestInstance<'a>

The test instance that was started.

TestFinished

A test finished running.

Fields of TestFinished

test_instance: TestInstance<'a>

The test instance that finished running.

run_status: TestRunStatus

Information about how this test was run.

TestSkipped

A test was skipped.

Fields of TestSkipped

test_instance: TestInstance<'a>

The test instance that was skipped.

RunBeginCancel

A cancellation notice was received.

Fields of RunBeginCancel

running: usize

The number of tests still running.

reason: CancelReason

The reason this run was canceled.

RunFinished

The test run finished.

Fields of RunFinished

start_time: Instant

The time at which the run was started.

initial_run_count: usize

The total number of tests that were expected to be run.

If the test run is canceled, this will be more than run_stats.tests_run.

run_stats: RunStats

Statistics for the run.